backend/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Redis
redis==5.0.1
hiredis==2.2.3

# AI Agent Frameworks
pyautogen==0.2.0
langgraph==0.0.28
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
crewai==0.1.0

# ML Models
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
imbalanced-learn==0.11.0

# Deep Learning
torch==2.1.1
torch-geometric==2.4.0

# Transformers & NLP
transformers==4.36.0
sentence-transformers==2.2.2
tokenizers==0.15.0

# Explainability
shap==0.43.0
lime==0.2.0.1

# Vector Database
pinecone-client==3.0.0
openai==1.6.0

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Monitoring & Observability
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk[fastapi]==1.39.1

# Security
cryptography==41.0.7
python-dotenv==1.0.0

# Payment Processing
stripe==7.8.0

# Background Jobs
celery[redis]==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
